% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AdjBQR.R
\name{li_reg}
\alias{li_reg}
\title{Asymmetric Laplace Working Likelihood For
Linear Quantile Regression}
\usage{
li_reg(pars, y, x, tau, sig)
}
\arguments{
\item{pars}{regression coefficient vector}

\item{y}{the response vector}

\item{x}{the design matrix with one in the first column corresponding to the intercept}

\item{tau}{the quantile level}

\item{sig}{scale parameter sigma}
}
\value{
the working log (asymmetric Laplace) likelihood function (the part involving the regression coefficients)
}
\description{
Asymmetric Laplace Working Likelihood For
Linear Quantile Regression
}
\details{
The asymmetric Laplace working likelihood
 is proportional to exponential
of the negative quantile objective function for
linear quantile regression
}
\references{
Yang, Y., Wang, H. and He, X. (2015). Posterior inference in Bayesian quantile regression with asymmetric Laplace
  likelihood. International Statistical Review, 2015. doi: 10.1111/insr.12114.
}

