% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AdjBQR.R
\name{sigma_est_powell}
\alias{sigma_est_powell}
\title{Estimation of the Scale Parameter for Censored Quantile Regression}
\usage{
sigma_est_powell(y, x)
}
\arguments{
\item{y}{the response vector}

\item{x}{the design matrix with one in the first column corresponding to the intercept}
}
\value{
the estimated scale parameter sigma
}
\description{
Estimation of the Scale Parameter for Censored Quantile Regression
}

